Stop running TravisCI against jRuby for now
I am starting to find it difficult to argue in favor of having jruby in
the list of TravisCI rubies.

It adds a bit of extra cost in maintenance and time spent waiting build to finish,
without feeling we're getting a lot out of it. Thus, the feeling is that it has low ROI.

Reasons behind my feeling of not "getting a lot out of it" includes:

- Almost never coming across a situation in which I thought to myself
"Hey, with this change we're making the gem incompatible with jRuby" because most
of the failures on jRuby builds where either heisenbugs and/or rvm installation
problems with jRuby, at least in my experience

- Usage share seems to be very very low, even compared to unmaintaned
MRI version, according to some sources, e.g.
https://semaphoreci.com/blog/2017/11/08/ruby-versions-used-in-commercial-projects-in-2017.html