Rescue InvalidParameterError with 400 Response
The documentation for [Rack::QueryParser#parse_nested_query][] states
that when encountering an `InvalidParameterError`, the middleware or application
should just return an empty 400 response to indicate that the request will not be
processed, rather than allowing the exception to throw a 500 error. This
happens in [Rails][], so when a particular request/IP combo is being
thwarted by `Rack::Attack`, you'll still get an error that ends up being
a bunch of noise in your error monitoring or logging system. To address
this, `rescue` the errors that could cause this particular situation and
respond with a clean 400 error.

[Rack::QueryParser#parse_nested_query]: https://www.rubydoc.info/github/rack/rack/Rack%2FQueryParser:parse_nested_query
[Rails]: https://github.com/rails/rails/blob/fdd5894aa985651fb99e0ef1c2d3f3f604b2f659/actionpack/lib/action_dispatch/middleware/exception_wrapper.rb#L23